rust: use serde for the TileMux interface.

+++ src/apps/bench/rustbenchs/src/bipc.rs
@@ -101,7 +101,7 @@ fn pingpong_with_multiple(t: &mut dyn WvTester) {
     // use long time slices for childs (minimize timer interrupts)
     wv_assert_ok!(tile.set_quota(
         TimeDuration::from_secs(1).as_raw(),
-        tile.quota().unwrap().page_tables().left() as u64,
+        tile.quota().unwrap().page_tables().left(),
     ));
 
     // split time quota between childs
